#2 Base Building - Basic Commander Cam function
Functionality providing the builder a top-down view of the base area.
Can be in future limited to a base area which will be defined in future.
Needs to be wired up to the future Commander system to allow the
commander to open the build menu from within and from the surrounds of
the HQ.